refactor(cloudy): rename mirage eDSL package internal.edsl to edsl to match its public authoring surface
