+++ cloudy/src/commonMain/kotlin/com/skydoves/cloudy/MirageParams.kt
@@ -23,13 +23,13 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.ImageBitmap
 import androidx.compose.ui.graphics.TileMode
 import com.skydoves.cloudy.internal.UniformEntry
-import com.skydoves.cloudy.internal.edsl.Expression
-import com.skydoves.cloudy.internal.edsl.Float1
-import com.skydoves.cloudy.internal.edsl.Float2
-import com.skydoves.cloudy.internal.edsl.Float4
-import com.skydoves.cloudy.internal.edsl.Half4
-import com.skydoves.cloudy.internal.edsl.ShaderType
-import com.skydoves.cloudy.internal.edsl.UniformRef
+import com.skydoves.cloudy.edsl.Expression
+import com.skydoves.cloudy.edsl.Float1
+import com.skydoves.cloudy.edsl.Float2
+import com.skydoves.cloudy.edsl.Float4
+import com.skydoves.cloudy.edsl.Half4
+import com.skydoves.cloudy.edsl.ShaderType
+import com.skydoves.cloudy.edsl.UniformRef
 import kotlin.properties.PropertyDelegateProvider
 import kotlin.properties.ReadOnlyProperty
 import kotlin.reflect.KProperty
