+++ cloudy/src/commonMain/kotlin/com/skydoves/cloudy/MirageShaders.kt
@@ -20,57 +20,57 @@ package com.skydoves.cloudy
 import androidx.compose.ui.geometry.Offset
 import androidx.compose.ui.geometry.Size
 import androidx.compose.ui.graphics.Color
-import com.skydoves.cloudy.internal.edsl.Float2
-import com.skydoves.cloudy.internal.edsl.Half4
-import com.skydoves.cloudy.internal.edsl.If
-import com.skydoves.cloudy.internal.edsl.a
-import com.skydoves.cloudy.internal.edsl.abs
-import com.skydoves.cloudy.internal.edsl.and
-import com.skydoves.cloudy.internal.edsl.boxRoundedSDF
-import com.skydoves.cloudy.internal.edsl.clamp
-import com.skydoves.cloudy.internal.edsl.cos
-import com.skydoves.cloudy.internal.edsl.div
-import com.skydoves.cloudy.internal.edsl.dot
-import com.skydoves.cloudy.internal.edsl.exp
-import com.skydoves.cloudy.internal.edsl.float1
-import com.skydoves.cloudy.internal.edsl.float2
-import com.skydoves.cloudy.internal.edsl.float3
-import com.skydoves.cloudy.internal.edsl.float4
-import com.skydoves.cloudy.internal.edsl.floor
-import com.skydoves.cloudy.internal.edsl.foilHash
-import com.skydoves.cloudy.internal.edsl.fract
-import com.skydoves.cloudy.internal.edsl.greaterThan
-import com.skydoves.cloudy.internal.edsl.greaterThanEqual
-import com.skydoves.cloudy.internal.edsl.guard
-import com.skydoves.cloudy.internal.edsl.half
-import com.skydoves.cloudy.internal.edsl.half3
-import com.skydoves.cloudy.internal.edsl.half4
-import com.skydoves.cloudy.internal.edsl.length
-import com.skydoves.cloudy.internal.edsl.lensNormalDirection
-import com.skydoves.cloudy.internal.edsl.lessThanEqual
-import com.skydoves.cloudy.internal.edsl.local
-import com.skydoves.cloudy.internal.edsl.luma
-import com.skydoves.cloudy.internal.edsl.max
-import com.skydoves.cloudy.internal.edsl.min
-import com.skydoves.cloudy.internal.edsl.minus
-import com.skydoves.cloudy.internal.edsl.mirageTime
-import com.skydoves.cloudy.internal.edsl.mix
-import com.skydoves.cloudy.internal.edsl.normalize
-import com.skydoves.cloudy.internal.edsl.plus
-import com.skydoves.cloudy.internal.edsl.pow
-import com.skydoves.cloudy.internal.edsl.processColor
-import com.skydoves.cloudy.internal.edsl.rgb
-import com.skydoves.cloudy.internal.edsl.sampleContent
-import com.skydoves.cloudy.internal.edsl.signSelect
-import com.skydoves.cloudy.internal.edsl.sin
-import com.skydoves.cloudy.internal.edsl.smoothstep
-import com.skydoves.cloudy.internal.edsl.sqrt
-import com.skydoves.cloudy.internal.edsl.step
-import com.skydoves.cloudy.internal.edsl.times
-import com.skydoves.cloudy.internal.edsl.unaryMinus
-import com.skydoves.cloudy.internal.edsl.x
-import com.skydoves.cloudy.internal.edsl.xyz
-import com.skydoves.cloudy.internal.edsl.y
+import com.skydoves.cloudy.edsl.Float2
+import com.skydoves.cloudy.edsl.Half4
+import com.skydoves.cloudy.edsl.If
+import com.skydoves.cloudy.edsl.a
+import com.skydoves.cloudy.edsl.abs
+import com.skydoves.cloudy.edsl.and
+import com.skydoves.cloudy.edsl.boxRoundedSDF
+import com.skydoves.cloudy.edsl.clamp
+import com.skydoves.cloudy.edsl.cos
+import com.skydoves.cloudy.edsl.div
+import com.skydoves.cloudy.edsl.dot
+import com.skydoves.cloudy.edsl.exp
+import com.skydoves.cloudy.edsl.float1
+import com.skydoves.cloudy.edsl.float2
+import com.skydoves.cloudy.edsl.float3
+import com.skydoves.cloudy.edsl.float4
+import com.skydoves.cloudy.edsl.floor
+import com.skydoves.cloudy.edsl.foilHash
+import com.skydoves.cloudy.edsl.fract
+import com.skydoves.cloudy.edsl.greaterThan
+import com.skydoves.cloudy.edsl.greaterThanEqual
+import com.skydoves.cloudy.edsl.guard
+import com.skydoves.cloudy.edsl.half
+import com.skydoves.cloudy.edsl.half3
+import com.skydoves.cloudy.edsl.half4
+import com.skydoves.cloudy.edsl.length
+import com.skydoves.cloudy.edsl.lensNormalDirection
+import com.skydoves.cloudy.edsl.lessThanEqual
+import com.skydoves.cloudy.edsl.local
+import com.skydoves.cloudy.edsl.luma
+import com.skydoves.cloudy.edsl.max
+import com.skydoves.cloudy.edsl.min
+import com.skydoves.cloudy.edsl.minus
+import com.skydoves.cloudy.edsl.mirageTime
+import com.skydoves.cloudy.edsl.mix
+import com.skydoves.cloudy.edsl.normalize
+import com.skydoves.cloudy.edsl.plus
+import com.skydoves.cloudy.edsl.pow
+import com.skydoves.cloudy.edsl.processColor
+import com.skydoves.cloudy.edsl.rgb
+import com.skydoves.cloudy.edsl.sampleContent
+import com.skydoves.cloudy.edsl.signSelect
+import com.skydoves.cloudy.edsl.sin
+import com.skydoves.cloudy.edsl.smoothstep
+import com.skydoves.cloudy.edsl.sqrt
+import com.skydoves.cloudy.edsl.step
+import com.skydoves.cloudy.edsl.times
+import com.skydoves.cloudy.edsl.unaryMinus
+import com.skydoves.cloudy.edsl.x
+import com.skydoves.cloudy.edsl.xyz
+import com.skydoves.cloudy.edsl.y
 
 /**
  * Bundled [MirageShader] presets — the catalog of ready-to-apply looks.
