+++ cloudy/src/commonMain/kotlin/com/skydoves/cloudy/RainyWindowShader.kt
@@ -18,53 +18,53 @@
 package com.skydoves.cloudy
 
 import androidx.compose.ui.graphics.TileMode
-import com.skydoves.cloudy.internal.edsl.Argument
-import com.skydoves.cloudy.internal.edsl.Float1
-import com.skydoves.cloudy.internal.edsl.Float2
-import com.skydoves.cloudy.internal.edsl.Float3
-import com.skydoves.cloudy.internal.edsl.Half4
-import com.skydoves.cloudy.internal.edsl.If
-import com.skydoves.cloudy.internal.edsl.ShaderType
-import com.skydoves.cloudy.internal.edsl.a
-import com.skydoves.cloudy.internal.edsl.abs
-import com.skydoves.cloudy.internal.edsl.clamp
-import com.skydoves.cloudy.internal.edsl.defineHelper
-import com.skydoves.cloudy.internal.edsl.div
-import com.skydoves.cloudy.internal.edsl.dot
-import com.skydoves.cloudy.internal.edsl.eval
-import com.skydoves.cloudy.internal.edsl.float1
-import com.skydoves.cloudy.internal.edsl.float2
-import com.skydoves.cloudy.internal.edsl.float3
-import com.skydoves.cloudy.internal.edsl.floor
-import com.skydoves.cloudy.internal.edsl.fract
-import com.skydoves.cloudy.internal.edsl.greaterThan
-import com.skydoves.cloudy.internal.edsl.half
-import com.skydoves.cloudy.internal.edsl.half3
-import com.skydoves.cloudy.internal.edsl.half4
-import com.skydoves.cloudy.internal.edsl.length
-import com.skydoves.cloudy.internal.edsl.local
-import com.skydoves.cloudy.internal.edsl.max
-import com.skydoves.cloudy.internal.edsl.min
-import com.skydoves.cloudy.internal.edsl.minus
-import com.skydoves.cloudy.internal.edsl.mirageResolution
-import com.skydoves.cloudy.internal.edsl.mirageTime
-import com.skydoves.cloudy.internal.edsl.mix
-import com.skydoves.cloudy.internal.edsl.mod
-import com.skydoves.cloudy.internal.edsl.plus
-import com.skydoves.cloudy.internal.edsl.r
-import com.skydoves.cloudy.internal.edsl.rgb
-import com.skydoves.cloudy.internal.edsl.sampleContent
-import com.skydoves.cloudy.internal.edsl.sin
-import com.skydoves.cloudy.internal.edsl.smoothstep
-import com.skydoves.cloudy.internal.edsl.sqrt
-import com.skydoves.cloudy.internal.edsl.times
-import com.skydoves.cloudy.internal.edsl.unaryMinus
-import com.skydoves.cloudy.internal.edsl.x
-import com.skydoves.cloudy.internal.edsl.xy
-import com.skydoves.cloudy.internal.edsl.y
-import com.skydoves.cloudy.internal.edsl.yx
-import com.skydoves.cloudy.internal.edsl.yzx
-import com.skydoves.cloudy.internal.edsl.z
+import com.skydoves.cloudy.edsl.Argument
+import com.skydoves.cloudy.edsl.Float1
+import com.skydoves.cloudy.edsl.Float2
+import com.skydoves.cloudy.edsl.Float3
+import com.skydoves.cloudy.edsl.Half4
+import com.skydoves.cloudy.edsl.If
+import com.skydoves.cloudy.edsl.ShaderType
+import com.skydoves.cloudy.edsl.a
+import com.skydoves.cloudy.edsl.abs
+import com.skydoves.cloudy.edsl.clamp
+import com.skydoves.cloudy.edsl.defineHelper
+import com.skydoves.cloudy.edsl.div
+import com.skydoves.cloudy.edsl.dot
+import com.skydoves.cloudy.edsl.eval
+import com.skydoves.cloudy.edsl.float1
+import com.skydoves.cloudy.edsl.float2
+import com.skydoves.cloudy.edsl.float3
+import com.skydoves.cloudy.edsl.floor
+import com.skydoves.cloudy.edsl.fract
+import com.skydoves.cloudy.edsl.greaterThan
+import com.skydoves.cloudy.edsl.half
+import com.skydoves.cloudy.edsl.half3
+import com.skydoves.cloudy.edsl.half4
+import com.skydoves.cloudy.edsl.length
+import com.skydoves.cloudy.edsl.local
+import com.skydoves.cloudy.edsl.max
+import com.skydoves.cloudy.edsl.min
+import com.skydoves.cloudy.edsl.minus
+import com.skydoves.cloudy.edsl.mirageResolution
+import com.skydoves.cloudy.edsl.mirageTime
+import com.skydoves.cloudy.edsl.mix
+import com.skydoves.cloudy.edsl.mod
+import com.skydoves.cloudy.edsl.plus
+import com.skydoves.cloudy.edsl.r
+import com.skydoves.cloudy.edsl.rgb
+import com.skydoves.cloudy.edsl.sampleContent
+import com.skydoves.cloudy.edsl.sin
+import com.skydoves.cloudy.edsl.smoothstep
+import com.skydoves.cloudy.edsl.sqrt
+import com.skydoves.cloudy.edsl.times
+import com.skydoves.cloudy.edsl.unaryMinus
+import com.skydoves.cloudy.edsl.x
+import com.skydoves.cloudy.edsl.xy
+import com.skydoves.cloudy.edsl.y
+import com.skydoves.cloudy.edsl.yx
+import com.skydoves.cloudy.edsl.yzx
+import com.skydoves.cloudy.edsl.z
 
 /**
  * Rain on steamed glass: running refracting raindrops + condensation you wipe clear with a finger.
