+++ cloudy/src/desktopTest/kotlin/com/skydoves/cloudy/edsl/ShaderLiteralsTest.kt
@@ -15,7 +15,7 @@
  */
 @file:OptIn(ExperimentalMirage::class)
 
-package com.skydoves.cloudy.internal.edsl
+package com.skydoves.cloudy.edsl
 
 import androidx.compose.ui.geometry.Offset
 import androidx.compose.ui.geometry.Size
